Refactor greeting-related components into a dedicated package.
Moved Greeting, Platform, and related logic to a new "greeting" package to improve code organization and maintainability. Extracted `GreetingView` into a separate composable function for better component reuse. Updated imports and package paths across affected files accordingly.
