+++ composeApp/src/commonMain/kotlin/ch/rrte/pl/tasky/greeting/Greeting.kt
@@ -1,4 +1,4 @@
-package ch.rrte.pl.tasky
+package ch.rrte.pl.tasky.greeting
 
 class Greeting {
     private val platform = getPlatform()
