+++ composeApp/src/commonMain/kotlin/ch/rrte/pl/tasky/greeting/Platform.kt
@@ -1,4 +1,4 @@
-package ch.rrte.pl.tasky
+package ch.rrte.pl.tasky.greeting
 
 interface Platform {
     val name: String
